requirements.txt
asgiref==3.5.0
cffi==1.15.0
cryptography==36.0.2
Django==3.2
django-cors-headers==3.11.0
django-filter==21.1
django-mssql-backend==2.8.1
django-password-reset==2.0
django-pyodbc==1.1.3
django-reset-migrations==0.4.0
django-rest-knox==4.2.0
django-rest-passwordreset==1.2.1
djangorestframework==3.13.1
gunicorn==20.1.0
importlib-metadata==4.11.3
Markdown==3.3.6
pycparser==2.21
pyodbc==4.0.32
pytz==2022.1
rest-framework-pagination==0.0.1
sql-server.pyodbc==1.0
sqlparse==0.4.2
zipp==3.7.0
